back/pyproject.toml
[tool.poetry]
name = "employee state monitoring"
version = "0.1.0"
description = ""
authors = ["vsevalot <v.v.tetervak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.1"
uvicorn = "^0.23.1"
fire = "^0.5.0"
sqlalchemy = "^2.0.19"
alembic = "^1.11.1"
pydantic-settings = "^2.0.2"
asyncpg = "^0.28.0"
greenlet = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
